refactor - bring component using only rest to main_rest_container
